splitting the utils.py file into two separate files
